Move animal nodes
Move the animal nodes from the tests package to the testing helpers
package.

These nodes are used by tests in order to not depend on the nodes
defined in Arborista because doing so is brittle. Id est, nodes defined
by Abrorista are likely to change which would necessitate also changing
the test cases. Having to change the test cases in this scenario is not
desirable. In an ideal world only tests which test the unit of code
being changed should have to change.

There are still a lot of tests which use Arborista nodes that should
eventually be converted to use testing defined nodes. We probably need
a better testing node set though in order for transformations to make
sense.

One school of thought is that we should try to make the testing nodes
refer to items which are common knowledge. This would help make test
cases more understandable.

Regardless of all of these todos, moving the animal nodes into the
testing helpers makes more sense than having them in the tests package
because the animal nodes themselves are not tests. It might also make
sense to have a seperate testing data package. Or, down the road, have
a mechanism for dynamically creating desired nodes, trees, and
transformations from a string (a domain specific language).
